requirements.txt
django
djangorestframework
marshmallow
django-requestlogs